requirements.txt
# Core Streamlit and Web Framework
streamlit==1.29.0
streamlit-webrtc==0.47.1

# AssemblyAI SDK and Audio Processing (Updated for Universal-Streaming)
assemblyai>=0.30.0
websockets==12.0
aiohttp==3.9.1

# Audio Processing and Real-time Streaming (Windows-friendly)
# pyaudio==0.2.11  # Commented out - difficult to install on Windows
sounddevice==0.4.6  # Better cross-platform audio library
numpy==1.24.3
scipy==1.11.4 
librosa==0.10.1
soundfile==0.12.1
# webrtcvad==2.0.10  # Optional - can cause issues on Windows

# Environment and Configuration Management
python-dotenv==1.0.0
pydantic==2.5.2
pydantic-settings==2.1.0

# Data Processing and Analysis
pandas==2.1.4
plotly==5.17.0
altair==5.2.0

# Async and Real-time Processing
# asyncio-mqtt==0.16.1  # Optional
# asyncio==3.4.3  # Built into Python 3.7+
threading-timer==0.1.0

# UI Enhancement and Styling
streamlit-option-menu==0.3.6
streamlit-elements==0.1.0
streamlit-aggrid==0.3.4.post3
streamlit-lottie==0.0.5

# Performance and Monitoring
psutil==5.9.6
memory-profiler==0.61.0
line-profiler==4.1.1

# Logging and Error Handling
loguru==0.7.2
sentry-sdk==1.39.2

# Date and Time Handling
python-dateutil==2.8.2
pytz==2023.3

# HTTP Requests and API Integration
requests==2.31.0
httpx==0.25.2

# JSON and Data Serialization
orjson==3.9.10
msgpack==1.0.7

# Security and Validation
cryptography==41.0.8
validators==0.22.0

# Testing and Development (Optional)
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0

# Additional Utilities
click==8.1.7
tqdm==4.66.1
colorama==0.4.6